Дирректория coliving_common/ вынесен общий код
